requirements.txt
asgiref==3.8.1
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.2
sqlparse==0.5.3
tzdata==2024.2
